+++ src/Controller/Admin/BackupsController.php
@@ -25,7 +25,6 @@ namespace MeCms\Controller\Admin;
 use Cake\Cache\Cache;
 use Cake\Core\Configure;
 use MeCms\Controller\AppController;
-use MeCms\Controller\Traits\DownloadTrait;
 use MysqlBackup\Utility\BackupImport;
 use MysqlBackup\Utility\BackupManager;
 
@@ -34,8 +33,6 @@ use MysqlBackup\Utility\BackupManager;
  */
 class BackupsController extends AppController
 {
-    use DownloadTrait;
-
     /**
      * Check if the provided user is authorized for the request
      * @param array $user The user to check the authorization of. If empty
@@ -130,11 +127,12 @@ class BackupsController extends AppController
      * Downloads a backup file
      * @param string $filename Backup filename
      * @return \Cake\Network\Response
-     * @uses MeCms\Controller\Traits\DownloadTrait::_download()
      */
     public function download($filename)
     {
-        return $this->_download(Configure::read('MysqlBackup.target') . DS . urldecode($filename));
+        $file = Configure::read('MysqlBackup.target') . DS . urldecode($filename);
+
+        return $this->response->withFile($file, ['download' => true]);
     }
 
     /**